+++ include/CppLinuxSerial/SerialPort.hpp
@@ -2,7 +2,7 @@
 /// \file 			SerialPort.hpp
 /// \author 		Geoffrey Hunter <gbmhunter@gmail.com> (www.mbedded.ninja)
 /// \created		2014-01-07
-/// \last-modified 	2022-11-12
+/// \last-modified 	2023-02-13
 /// \brief			The main serial port class.
 /// \details
 ///					See README.rst in repo root dir for more info.
@@ -171,20 +171,16 @@ namespace mn {
             /// \throws     CppLinuxSerial::Exception if state != OPEN.
             void WriteBinary(const std::vector<uint8_t>& data);
 
-            /// \brief      Use to read text from the COM port.
-            /// \param      data        The object the read characters from the COM port will be saved to.
-            /// \param      wait_ms     The amount of time to wait for data. Set to 0 for non-blocking mode. Set to -1
-            ///                 to wait indefinitely for new data.
+            /// \brief      Use to read text from the COM port. Blocking nature depends on SetTimeout().
+            /// \param      data        The read characters from the COM port will be appended to this string.
             /// \note       Use ReadBinary() if you want to interpret received data as binary.
             /// \throws     
             ///             CppLinuxSerial::Exception if state != OPEN.
             ///             std::system_error() if device has been disconnected.
             void Read(std::string& data);
 
-            /// \brief      Use to read binary data from the COM port.
-            /// \param      data        The object the read uint8_t bytes from the COM port will be saved to.
-            /// \param      wait_ms     The amount of time to wait for data. Set to 0 for non-blocking mode. Set to -1
-            ///                 to wait indefinitely for new data.
+            /// \brief      Use to read binary data from the COM port. Blocking nature depends on SetTimeout().
+            /// \param      data        The read bytes from the COM port will be appended to this vector.
             /// \note       Use Read() if you want to interpret received data as a string.
             /// \throws     CppLinuxSerial::Exception if state != OPEN.
             ///             std::system_error() if device has been disconnected.